Split enclave manager to base and singleton
- Carve out base_enclave_manager from enclave_manger to
  create base class EnclaveManager
- Inherit from BaseEnclaveManager in SingletonEnclaveManager
